% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{extract_about_links}
\alias{extract_about_links}
\title{Extract links and other information related to about page}
\usage{
extract_about_links(base_url)
}
\arguments{
\item{base_url}{A base URL (the base part of the web address)}
}
\value{
If successful, the function returns a dataframe of three columns ('href', 'link_text', link'). If not, the function reports one of the following five error cases: "Found without tree search.", "This website is broken.", "The website is flat (no tree structure).", "PHP error", or "The website does not have about page."
}
\description{
Extract links and other information related to about page
}
